增加rollback功能

+++ wrap/business.py
@@ -64,15 +64,31 @@ def backup(mod):
 
 def clean(mod):
     exp = '%s/%s-*.tar' % (mod.bakdir(),mod.name())
-    print exp
     for ip in mod.deploy():
         cmd = 'ls %s | sort | tail -5; ls %s | wc -w' % (exp,exp)
         remoteCmd(ip, cmd)
         instr = raw_input('请输入要清理的数量: ')
+        if instr == '':
+            break
         if instr.isdigit() == False:
             continue
         num = int(instr)
         cmd = 'rm -rf `ls %s | sort | head -%d`' % (exp,num)
+        out = ask('将在 (%s) 运行命令 (%s), 确认立刻执行吗？' % (ip,cmd), 'yes,no', 'no')
+        if out == 'yes':
+            remoteCmd(ip, cmd)
+
+def rollback(mod):
+    exp = '%s/%s-*.tar' % (mod.bakdir(),mod.name())
+    for ip in mod.deploy():
+        cmd = 'ls %s | sort | tail -5' % exp
+        remoteCmd(ip, cmd)
+        instr = raw_input('请输入要回滚的备份文件的时间戳: ')
+        if instr == '':
+            break
+        if instr.isdigit() == False:
+            continue
+        cmd = 'rm -rf %s && tar -xf %s/%s-%s.tar -C %s' % (mod.appdir(),mod.bakdir(),mod.name(),instr,mod.upappdir())
         out = ask('将在 (%s) 运行命令 (%s), 确认立刻执行吗？' % (ip,cmd), 'yes,no', 'no')
         if out == 'yes':
             remoteCmd(ip, cmd)
